+++ src/modules/product/product.service.ts
@@ -30,20 +30,52 @@ export class ProductService {
     return await this.productRepo.findOne(option);
   }
 
-  async update(id: string, dto: UpdateProductDto) {
-    const product = await this.findOne({ where: { id } });
+  async update(
+    id: string,
+    dto: UpdateProductDto,
+    pictures: Express.Multer.File[],
+  ) {
+    const product = await this.findOne({
+      where: { id },
+      relations: { images: true },
+    });
 
     if (!product) throw new NotFoundException('Product not found');
 
-    Object.assign(product, { ...dto });
+    Object.assign(product, dto);
 
     await this.productRepo.update(id, product);
+
+    if (product.images) {
+      for (const image of product.images) {
+        this.productImageService.remove(image.id);
+      }
+    }
+
+    this.productImageService.create({
+      files: pictures,
+      productId: product.id,
+    });
 
     return { message: 'Product Updated' };
   }
 
   async remove(id: string) {
-    await this.productRepo.delete(id);
+    const product = await this.findOne({
+      where: { id },
+      relations: { images: true },
+    });
+
+    if (!product) throw new NotFoundException('Product not found');
+
+    if (product.images) {
+      for (const image of product.images) {
+        this.productImageService.remove(image.id);
+      }
+    }
+
+    this.productRepo.delete(id);
+
     return { message: 'Product delete' };
   }
 }
